red: Cluster.Members wire shape + scanClusterRow member decode (#722) (tc-60kn)
